Simplify launch server
